+++ src/plugins/table-classnames/index.ts
@@ -1,0 +1,2 @@
+export { Config, create } from "./table-classnames";
+